WIP: Start adding PV plot functionality.
I actually wanted it to seem like a file in the tree view, but i
couldnt find a good way to manipulate the HDF5treeview.

Instead, I did override the createMenu function to add an PlotPV action.
This should in near future spawn a Plot2D that will get info from PV and
possibly restore it.