Fix PV restoring blinking in high acquisition rates.
In high acquisition rates, RestoreAction spent considerable time restoring the
pv image before plotting it, which resulted in a blinking image switching from
the raw image and the restored version.

To fix this, it is necessary that whenever the PVPlotter wants to replot a
restored image, it first restores the image and then plots it, instead
of plotting, capturing imageChanged signal and then restoring it and
plotting again.

I ended up understanding that since there are two different ways of controlling
processing of the array (one for the PVPlotter and other for the file reader), its ok
to make the RestoreAction two different classes: one for controlling the restoration in
file and other for controlling restoration in PV.