Unfortunate fix incoherent behavior on imageChanged signal
Some imageChanged signals could be triggered by events that are not actually changing
the data in image, but just metadata. We cannot restore image when this happens because
this would be a restoration of a restoration.

A better way to fix this would be: instead of creating class parameters intended to
store information about restoration state, make the Viewer reload the raw image and
restore it, similar to what is done in PVs. However, this would require a code refactor
beyond my time and priorization now...