Complete parametrization of array_prefix, width and height suffixes.
This commit finishes the main work being done in this branch.
Code is meant to get ArrayData PV preffix, width suffix and height
suffix from --prefix, --width-suffix and --height-suffix options.
If not found in options, search in defaults.yml file. If not found in file,
search in environment variable. If not found in environment variable, crash and cry.

Also, application will normally open file viewer normally, but
go straight to PVPlotter if --pv option (no arguments needed) is furnished.

Parser is only extending the parser created in silx_view_main.