Completely refactor (this should be more than just one commit)
I'm sorry for the long content of this commit.

The objective is to make adding a new specific transformation as easy as possible.
For that, developer only needs to add a function in the same format as pimega450DTrans or noTrans,
add the STATE associated with it in TransformToolButton.__init__, create an action associated with
the transformation in in TransformToolButton.createAllActions and add the action to the menu in this
same method.
Different icons can be added to different transformations.

TransforToolButton only warns the plot widgets when to actually transform things and
when to do nothing. It also warns them when to replot the raw image.

Plot widgets are responsible for keeping coherence between old and new images. Methods
do_transform() and recover() are there for that.
PVPlotter must have a different do_transform than FileViewer because if PVPlotter tries to use
sigActiveImageChanged to known when to plot transformed PV, it first plots the raw image, then
plots the transformation afterwards resulting in an epilepsy-crisis-generating flashing sequence of images.

recover can be the same method.