Add example_init file for deploy and new functionality.
example_init.sh plot a default PV when called with no arguments but
passes the user arguments if there are any.

I made this here because, as commented in the code, I still want the main
functionality of the application to plot files in an identical way to silx viewer,
but I also know beamline personnel will want to visualize PVs by default, depending
on the beamline.

Since I dont have an elegant way of putting that in python code yet, here we have another
quick-and-dirty bash script.